refactor(mm): split mapping area modules
